Reproducible research encourages devel. versioning
sessionInfo() will report 0.3.7 even if I have installed the development version of broom. This is bad for reproducible research and packages and scripts depending on the development version of broom.

See http://r-pkgs.had.co.nz/description.html#version

This commit bumps the version to 0.3.7.9000, so sessionInfor will report a version number easily identifiable as "development".

Feel free to reject this commit if you don't like this versioning scheme, personally I find it very useful and simple to apply.